Customer_Suport_Agents/requirements.txt
fastapi==0.116.0
uvicorn==0.25.0
langchain==0.1.0
langgraph==0.0.9
openai==1.31.0
faiss-cpu==1.13.2
python-dotenv==1.0.0
pydantic==2.9.0
sentence-transformers==2.3.1